@@ -39,6 +39,7 @@ export class AngularFireFunctions {
     const functions = of(undefined).pipe(
       observeOn(schedulers.outsideAngular),
       switchMap(() => import('firebase/functions')),
+      tap((it: any) => it),
       map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
       map(app => app.functions(region || undefined)),
       tap(functions => {
@@ -51,7 +52,7 @@ export class AngularFireFunctions {
 
     this.httpsCallable = <T = any, R = any>(name: string) =>
       (data: T) => from(functions).pipe(
-        observeOn(schedulers.outsideAngular),
+        observeOn(schedulers.insideAngular),
         switchMap(functions => functions.httpsCallable(name)(data)),
         map(r => r.data as R)
       );
