@@ -6,7 +6,10 @@ export function fromTask(task: UploadTask) {
     const progress = (snap: UploadTaskSnapshot) => subscriber.next(snap);
     const error = e => subscriber.error(e);
     const complete = () => subscriber.complete();
-    task.on('state_changed', progress, error, complete);
+    task.on('state_changed', progress, error, () => {
+      progress(task.snapshot);
+      complete();
+    });
     return () => task.cancel();
   });
 }
